restructured things
