+++ Fiesta-Server/src/api/skill.rs
@@ -1,31 +1,31 @@
+use crate::models::skill_model::Skill;
 use rocket::{response::status, serde::json::Json};
-
 // Mark skill as completed / not completed
 #[post("/c", format = "application/json", data = "<u>")]
-pub fn chk_skill(u: Json<Skill<'_>>) -> status::Accepted<String> {
+pub fn chk_skill(u: Json<Skill>) -> status::Accepted<String> {
     status::Accepted(Some(u.name))
 }
 
 // [ADMIN] - Create new Skill
 #[post("/p", format = "application/json", data = "<u>")]
-pub fn post_skill(u: Json<Skill<'_>>) -> status::Accepted<String> {
+pub fn post_skill(u: Json<Skill>) -> status::Accepted<String> {
     status::Accepted(Some(u.name))
 }
 
 // [ADMIN] - Delete Skill
 #[post("/d", format = "application/json", data = "<u>")]
-pub fn del_skill(u: Json<Skill<'_>>) -> status::Accepted<String> {
+pub fn del_skill(u: Json<Skill>) -> status::Accepted<String> {
     status::Accepted(Some(u.name))
 }
 
 // [ADMIN] - Create new Skill
 #[post("/n", format = "application/json", data = "<u>")]
-pub fn new_skill(u: Json<Skill<'_>>) -> status::Accepted<String> {
+pub fn new_skill(u: Json<Skill>) -> status::Accepted<String> {
     status::Accepted(Some(u.name))
 }
 
 // [ADMIN] - Create new Skill
 #[post("/u", format = "application/json", data = "<u>")]
-pub fn upd_skill(u: Json<Skill<'_>>) -> status::Accepted<String> {
+pub fn upd_skill(u: Json<Skill>) -> status::Accepted<String> {
     status::Accepted(Some(u.name))
 }
