+++ Fiesta-Server/src/models/skill_model.rs
@@ -1,6 +1,5 @@
 use mongodb::bson::oid::ObjectId;
 use serde::{Deserialize, Serialize};
-
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Skill {
     pub id: Option<ObjectId>,
