+++ Fiesta-Server/src/models/user_model.rs
@@ -1,10 +1,13 @@
 use mongodb::bson::oid::ObjectId;
 use serde::{Deserialize, Serialize};
 
+use super::skill_model::Skill;
+
 #[derive(Debug, Deserialize, Serialize)]
 pub struct User {
     pub id: Option<ObjectId>,
     pub name: String,
+    pub username: String,
     pub email: String,
     pub role: i32,                    // 0..2 -> 0:Admin, 1:Teacher, 2:Guest
     pub completed_skills: Vec<Skill>, //unaccessible if role > 1
